Removed the Thread.pass idle event.
This causes constant 100% CPU usage and isn't needed. On MRI threads *are*
executed but only after the main GTK event loop started (by the looks of it). On
Rubinius threads work just fine. Since Rubinius is my target Ruby I'm not really
going to bother with this idle loop hack.